Move some common widgets and utilities to `devtools_app_shared`
