+++ packages/devtools_app/lib/src/screens/deep_link_validation/project_root_selection/root_selector.dart
@@ -6,7 +6,6 @@ import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_app_shared/utils.dart';
 import 'package:flutter/material.dart';
 
-import '../../../shared/common_widgets.dart';
 import '../../../shared/primitives/utils.dart';
 import '../../../shared/ui/utils.dart';
