+++ packages/devtools_app/lib/src/shared/common_widgets.dart
@@ -14,7 +14,6 @@ import 'package:flutter/services.dart';
 import 'package:vm_service/vm_service.dart';
 
 import '../screens/debugger/debugger_controller.dart';
-import '../screens/inspector/layout_explorer/ui/theme.dart';
 import 'analytics/analytics.dart' as ga;
 import 'analytics/constants.dart' as gac;
 import 'config_specific/copy_to_clipboard/copy_to_clipboard.dart';
@@ -724,233 +723,6 @@ class InformationButton extends StatelessWidget {
       ),
     );
   }
-}
-
-class RoundedCornerOptions {
-  const RoundedCornerOptions({
-    this.showTopLeft = true,
-    this.showTopRight = true,
-    this.showBottomLeft = true,
-    this.showBottomRight = true,
-  });
-
-  /// Static constant instance with all borders hidden
-  static const empty = RoundedCornerOptions(
-    showTopLeft: false,
-    showTopRight: false,
-    showBottomLeft: false,
-    showBottomRight: false,
-  );
-
-  final bool showTopLeft;
-  final bool showTopRight;
-  final bool showBottomLeft;
-  final bool showBottomRight;
-}
-
-class RoundedDropDownButton<T> extends StatelessWidget {
-  const RoundedDropDownButton({
-    super.key,
-    this.value,
-    this.onChanged,
-    this.isDense = false,
-    this.isExpanded = false,
-    this.style,
-    this.selectedItemBuilder,
-    this.items,
-    this.roundedCornerOptions,
-  });
-
-  final T? value;
-
-  final ValueChanged<T?>? onChanged;
-
-  final bool isDense;
-
-  final bool isExpanded;
-
-  final TextStyle? style;
-
-  final DropdownButtonBuilder? selectedItemBuilder;
-
-  final List<DropdownMenuItem<T>>? items;
-
-  final RoundedCornerOptions? roundedCornerOptions;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final bgColor = theme.colorScheme.backgroundColorSelected;
-
-    Radius selectRadius(bool show) {
-      return show ? defaultRadius : Radius.zero;
-    }
-
-    final style = this.style ?? theme.regularTextStyle;
-    final showTopLeft = roundedCornerOptions?.showTopLeft ?? true;
-    final showTopRight = roundedCornerOptions?.showTopRight ?? true;
-    final showBottomLeft = roundedCornerOptions?.showBottomLeft ?? true;
-    final showBottomRight = roundedCornerOptions?.showBottomRight ?? true;
-
-    final button = Center(
-      child: SizedBox(
-        height: defaultButtonHeight - 2.0, // subtract 2.0 for width of border
-        child: DropdownButtonHideUnderline(
-          child: DropdownButton<T>(
-            padding: const EdgeInsets.only(
-              left: defaultSpacing,
-              right: borderPadding,
-            ),
-            value: value,
-            onChanged: onChanged,
-            isDense: isDense,
-            isExpanded: isExpanded,
-            borderRadius: BorderRadius.only(
-              topLeft: selectRadius(showTopLeft),
-              topRight: selectRadius(showTopRight),
-              bottomLeft: selectRadius(showBottomLeft),
-              bottomRight: selectRadius(showBottomRight),
-            ),
-            style: style,
-            selectedItemBuilder: selectedItemBuilder,
-            items: items,
-            focusColor: bgColor,
-          ),
-        ),
-      ),
-    );
-
-    if (roundedCornerOptions == RoundedCornerOptions.empty) return button;
-
-    return RoundedOutlinedBorder(
-      showTopLeft: showTopLeft,
-      showTopRight: showTopRight,
-      showBottomLeft: showBottomLeft,
-      showBottomRight: showBottomRight,
-      child: button,
-    );
-  }
-}
-
-class DevToolsClearableTextField extends StatelessWidget {
-  DevToolsClearableTextField({
-    super.key,
-    required this.labelText,
-    TextEditingController? controller,
-    this.hintText,
-    this.prefixIcon,
-    this.additionalSuffixActions = const <Widget>[],
-    this.onChanged,
-    this.onSubmitted,
-    this.autofocus = false,
-    this.enabled,
-    this.roundedBorder = false,
-  }) : controller = controller ?? TextEditingController();
-
-  final TextEditingController controller;
-  final String? hintText;
-  final Widget? prefixIcon;
-  final List<Widget> additionalSuffixActions;
-  final String labelText;
-  final void Function(String)? onChanged;
-  final void Function(String)? onSubmitted;
-  final bool autofocus;
-  final bool? enabled;
-  final bool roundedBorder;
-
-  static const _contentVerticalPadding = 6.0;
-
-  /// This is the default border radius used by the [OutlineInputBorder]
-  /// constructor.
-  static const _defaultInputBorderRadius =
-      BorderRadius.all(Radius.circular(4.0));
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return SizedBox(
-      height: defaultTextFieldHeight,
-      child: TextField(
-        autofocus: autofocus,
-        controller: controller,
-        enabled: enabled,
-        onChanged: onChanged,
-        onSubmitted: onSubmitted,
-        style: theme.regularTextStyle,
-        decoration: InputDecoration(
-          isDense: true,
-          contentPadding: const EdgeInsets.only(
-            top: _contentVerticalPadding,
-            bottom: _contentVerticalPadding,
-            left: denseSpacing,
-            right: densePadding,
-          ),
-          constraints: BoxConstraints(
-            minHeight: defaultTextFieldHeight,
-            maxHeight: defaultTextFieldHeight,
-          ),
-          border: OutlineInputBorder(
-            borderRadius: roundedBorder
-                ? const BorderRadius.all(defaultRadius)
-                : _defaultInputBorderRadius,
-          ),
-          labelText: labelText,
-          labelStyle: theme.subtleTextStyle,
-          hintText: hintText,
-          hintStyle: theme.subtleTextStyle,
-          prefixIcon: prefixIcon,
-          suffix: SizedBox(
-            height: inputDecorationElementHeight,
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                clearInputButton(
-                  () {
-                    controller.clear();
-                    onChanged?.call('');
-                  },
-                ),
-                ...additionalSuffixActions,
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-Widget clearInputButton(VoidCallback onPressed) {
-  return inputDecorationSuffixButton(
-    icon: Icons.clear,
-    onPressed: onPressed,
-    tooltip: 'Clear',
-  );
-}
-
-Widget closeSearchDropdownButton(VoidCallback? onPressed) {
-  return inputDecorationSuffixButton(icon: Icons.close, onPressed: onPressed);
-}
-
-Widget inputDecorationSuffixButton({
-  required IconData icon,
-  required VoidCallback? onPressed,
-  String? tooltip,
-}) {
-  return maybeWrapWithTooltip(
-    tooltip: tooltip,
-    child: SizedBox(
-      height: inputDecorationElementHeight,
-      width: inputDecorationElementHeight + denseSpacing,
-      child: IconButton(
-        padding: EdgeInsets.zero,
-        onPressed: onPressed,
-        iconSize: defaultIconSize,
-        splashRadius: defaultIconSize,
-        icon: Icon(icon),
-      ),
-    ),
-  );
 }
 
 class OutlinedRowGroup extends StatelessWidget {
