+++ packages/devtools_app/lib/src/shared/ui/drop_down_button.dart
@@ -7,7 +7,6 @@ import 'package:devtools_app_shared/ui.dart';
 import 'package:flutter/material.dart';
 
 import '../analytics/analytics.dart' as ga;
-import '../common_widgets.dart';
 
 /// A [DropDownButton] implementation that reports selection changes to our
 /// analytics.
