+++ packages/devtools_app_shared/lib/ui.dart
@@ -2,10 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+export 'src/ui/buttons.dart';
 export 'src/ui/common.dart';
 export 'src/ui/dialogs.dart';
 export 'src/ui/flex_split_column.dart';
 export 'src/ui/split_pane.dart';
+export 'src/ui/text_field.dart';
 export 'src/ui/theme/ide_theme.dart';
 export 'src/ui/theme/theme.dart';
 export 'src/ui/ui_utils.dart';
