+++ packages/devtools_app_shared/test/utils/utils_test.dart
@@ -94,4 +94,22 @@ void main() {
       expect(toCssHexColor(const Color(0xFFAABBCC)), equals('#aabbccff'));
     });
   });
+
+  group('StringExtension', () {
+    test('fuzzyMatch', () {
+      const str = 'hello_world_file';
+      expect(str.caseInsensitiveFuzzyMatch('h'), isTrue);
+      expect(str.caseInsensitiveFuzzyMatch('o_'), isTrue);
+      expect(str.caseInsensitiveFuzzyMatch('hw'), isTrue);
+      expect(str.caseInsensitiveFuzzyMatch('hwf'), isTrue);
+      expect(str.caseInsensitiveFuzzyMatch('_e'), isTrue);
+      expect(str.caseInsensitiveFuzzyMatch('HWF'), isTrue);
+      expect(str.caseInsensitiveFuzzyMatch('_E'), isTrue);
+
+      expect(str.caseInsensitiveFuzzyMatch('hwfh'), isFalse);
+      expect(str.caseInsensitiveFuzzyMatch('hfw'), isFalse);
+      expect(str.caseInsensitiveFuzzyMatch('gello'), isFalse);
+      expect(str.caseInsensitiveFuzzyMatch('files'), isFalse);
+    });
+  });
 }
